requirements.txt
aioredis==1.3.1
appdirs==1.4.3
asgiref==3.3.4
async-timeout==3.0.1
attrs==20.3.0
autobahn==21.3.1
Automat==20.2.0
CacheControl==0.12.6
certifi==2019.11.28
cffi==1.14.5
channels==3.0.3
channels-redis==3.2.0
chardet==3.0.4
colorama==0.4.3
constantly==15.1.0
contextlib2==0.6.0
cryptography==3.4.7
daphne==3.0.2
defusedxml==0.7.1
diff-match-patch==20200713
distlib==0.3.0
distro==1.4.0
Django==3.2
django-activity-stream==0.10.0
django-bulk-update-or-create==0.2.0
django-cors-headers==3.7.0
django-widget-tweaks==1.4.8
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
et-xmlfile==1.0.1
hiredis==2.0.0
html5lib==1.0.1
hyperlink==21.0.0
idna==3.1
incremental==21.3.0
ipaddr==2.2.0
lockfile==0.12.2
Markdown==3.3.4
MarkupPy==1.14
msgpack==1.0.2
numpy==1.20.2
odfpy==1.4.1
openpyxl==3.0.7
packaging==20.3
pandas==1.2.3
pep517==0.8.2
Pillow==8.2.0
progress==1.5
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==2.0.1
pyOpenSSL==20.0.1
pyparsing==2.4.6
python-dateutil==2.8.1
python-decouple==3.4
pytoml==0.1.21
pytz==2021.1
PyYAML==5.4.1
requests==2.22.0
retrying==1.3.3
service-identity==18.1.0
six==1.15.0
sqlparse==0.4.1
tablib==3.0.0
Twisted==21.2.0
txaio==21.2.1
urllib3==1.25.8
webencodings==0.5.1
websockets==8.1
xlrd==2.0.1
XlsxWriter==1.3.8
xlwt==1.3.0
zope.interface==5.3.0
